backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
motor>=3.6.0
pymupdf==1.24.4
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-groq==0.1.4
faiss-cpu==1.13.2
python-dotenv==1.0.1
python-multipart==0.0.9
sentence-transformers==2.7.0
aiofiles>=23.2.1
